+++ packages/crdt_lf_flutter/lib/crdt_lf_flutter.dart
@@ -6,11 +6,5 @@ library;
 export 'package:provider/provider.dart'
     show ProviderNotFoundException, ReadContext, SelectContext, WatchContext;
 
-export 'src/awareness_cursors.dart';
-export 'src/crdt_builder.dart';
-export 'src/crdt_handler.dart';
-export 'src/crdt_helper.dart';
-export 'src/crdt_provider.dart';
-export 'src/crdt_text_field.dart';
-export 'src/text_cursors.dart';
-export 'src/text_delta.dart';
+export 'src/effects/effects.dart';
+export 'src/provider/provider.dart';